third_party/blink/web_tests/external/wpt/notifications/lang.https-expected.txt

This is a testharness.js-based test.
[FAIL] Roundtripping lang "Latn-de". Expecting "".
  assert_equals: notification.lang expected "" but got "Latn-de"
[FAIL] Roundtripping lang "Latf-de". Expecting "".
  assert_equals: notification.lang expected "" but got "Latf-de"
[FAIL] Roundtripping lang "tic-tac-tac-toe". Expecting "".
  assert_equals: notification.lang expected "" but got "tic-tac-tac-toe"
[FAIL] Roundtripping lang "cocoa-1-bar". Expecting "".
  assert_equals: notification.lang expected "" but got "cocoa-1-bar"
[FAIL] Roundtripping lang "cocoa-a-bar". Expecting "".
  assert_equals: notification.lang expected "" but got "cocoa-a-bar"
[FAIL] Roundtripping lang "en-". Expecting "".
  assert_equals: notification.lang expected "" but got "en-"
[FAIL] Roundtripping lang "en--". Expecting "".
  assert_equals: notification.lang expected "" but got "en--"
[FAIL] Roundtripping lang "foo--bar". Expecting "".
  assert_equals: notification.lang expected "" but got "foo--bar"
[FAIL] Roundtripping lang "id---Java". Expecting "".
  assert_equals: notification.lang expected "" but got "id---Java"
[FAIL] Roundtripping lang "fr-x". Expecting "".
  assert_equals: notification.lang expected "" but got "fr-x"
[FAIL] Roundtripping lang "fr-xenomorph". Expecting "".
  assert_equals: notification.lang expected "" but got "fr-xenomorph"
[FAIL] Roundtripping lang "fr-x-xenomorph". Expecting "".
  assert_equals: notification.lang expected "" but got "fr-x-xenomorph"
[FAIL] Roundtripping lang "a". Expecting "".
  assert_equals: notification.lang expected "" but got "a"
[FAIL] Roundtripping lang "a-fr-lang". Expecting "".
  assert_equals: notification.lang expected "" but got "a-fr-lang"
[FAIL] Roundtripping lang "b-fr-lang". Expecting "".
  assert_equals: notification.lang expected "" but got "b-fr-lang"
[FAIL] Roundtripping lang "es1-KK-aa-bb-cc-dd". Expecting "".
  assert_equals: notification.lang expected "" but got "es1-KK-aa-bb-cc-dd"
[FAIL] Roundtripping lang "es2-KL-aa-bb-cc-dd". Expecting "".
  assert_equals: notification.lang expected "" but got "es2-KL-aa-bb-cc-dd"
[FAIL] Roundtripping lang "es3-KM-aa-bb-cc-dd". Expecting "".
  assert_equals: notification.lang expected "" but got "es3-KM-aa-bb-cc-dd"
[FAIL] Roundtripping lang "fooÉ". Expecting "".
  assert_equals: notification.lang expected "" but got "fooÉ"
[FAIL] Roundtripping lang "foöÉ-bÁr". Expecting "".
  assert_equals: notification.lang expected "" but got "foöÉ-bÁr"
[FAIL] Roundtripping lang "foöÉbÁr". Expecting "".
  assert_equals: notification.lang expected "" but got "foöÉbÁr"
Harness: the test ran to completion.
